# docs/swagger.yaml
definitions:
  domain.BaseUser:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    required:
    - email
    - id
    - username
    type: object
  domain.Comment:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      task_id:
        type: string
      user:
        $ref: '#/definitions/domain.BaseUser'
    required:
    - comment
    - created_at
    - id
    - task_id
    - user
    type: object
  domain.FindTasksResponse:
    properties:
      tasks:
        items:
          $ref: '#/definitions/domain.Task'
        type: array
      total:
        type: integer
    required:
    - tasks
    - total
    type: object
...
  domain.History:
    properties:
      changed_at:
        type: string
      changed_by:
        $ref: '#/definitions/domain.BaseUser'
      id:
        type: string
      snapshot:
        $ref: '#/definitions/domain.Task'
      task_id:
        type: string
    required:
    - changed_at
    - changed_by
    - id
    - snapshot
    - task_id
    type: object
  domain.IDResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
    required:
    - id
    type: object
...
  domain.Status:
    enum:
    - todo
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - StatusTodo
    - StatusInProgress
    - StatusDone
  domain.Task:
    properties:
      assignee:
        $ref: '#/definitions/domain.BaseUser'
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      owner:
        $ref: '#/definitions/domain.BaseUser'
      status:
        $ref: '#/definitions/domain.Status'
      team_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - created_at
    - id
    - owner
    - status
    - team_id
    - title
    - updated_at
    type: object
...
  domain.ValidatationErrors:
    properties:
      errors:
        items:
          $ref: '#/definitions/domain.ValidationError'
        type: array
    type: object
  domain.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  httpcommon.MessageResponse:
    properties:
      message:
        type: string
    type: object
  inputs.CreateComment:
    properties:
      text:
        example: Комментарий к задаче
        maxLength: 500
        minLength: 1
        type: string
    required:
    - text
    type: object
  inputs.CreateTask:
    properties:
      assignee_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
      description:
        example: Description of the task
        type: string
      team_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
      title:
        example: Task title
        type: string
    required:
    - team_id
    - title
    type: object
...
  inputs.UpdateTask:
    properties:
      description:
        example: Description of the task
        type: string
      status:
        $ref: '#/definitions/domain.Status'
      title:
        example: Task title
        type: string
    required:
    - title
    type: object
info:
  contact:
    email: f3d0t0tvmax@yandex.ru
    name: Fedotv Maxim (developer)
  description: Swagger Documentation for mkk_luna_test_rest_api (тестовое задание
    для компании "МКК ЛУНА")
  title: Swagger Documentation for mkk_luna_test_rest_api
  version: "1.0"
paths:
...
  /api/v1/tasks:
    get:
      consumes:
      - application/json
      description: Получить список задач команды с фильтрацией и пагинацией. Доступно
        любому участнику команды
      parameters:
      - description: Team ID
        format: uuid
        in: query
        name: team_id
        required: true
        type: string
      - description: Статус задачи
        enum:
        - todo
        - in_progress
        - done
        in: query
        name: status
        type: string
      - description: ID исполнителя
        format: uuid
        in: query
        name: assignee_id
        type: string
      - description: Номер страницы
        example: 1
        in: query
        name: page
        type: integer
      - description: Размер страницы
        example: 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.FindTasksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Получить список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать задачу в команде, пользователь должен быть членом команды
      parameters:
      - description: Объект для создания задачи
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/inputs.CreateTask'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Создать задачу
      tags:
      - tasks
  /api/v1/tasks/{id}:
    put:
      consumes:
      - application/json
      description: Обновить задачу. Пользователь должен иметь права (owner/admin/assignee)
      parameters:
      - description: Task ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Объект для обновления задачи
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/inputs.UpdateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Обновить задачу
      tags:
      - tasks
  /api/v1/tasks/{id}/comments:
    get:
      consumes:
      - application/json
      description: Получить список комментариев задачи. Доступно любому участнику
        команды
      parameters:
      - description: Task ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Получить комментарии задачи
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать комментарий к задаче. Пользователь должен быть членом команды
        задачи
      parameters:
      - description: Task ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Объект для создания комментария
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/inputs.CreateComment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Создать комментарий к задаче
      tags:
      - tasks
  /api/v1/tasks/{id}/history:
    get:
      consumes:
      - application/json
      description: Получить историю изменений задачи. Доступно любому участнику команды
      parameters:
      - description: Task ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.History'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Получить историю изменений задачи
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    description: Введите "Bearer [ваш JWT токен]"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"